Added Network manager for communication with Raspberry Pi server. New Stations manager for checking stations connections. Charts and Database improvements.

+++ iOS/WeatherStation/WeatherStation/Measurements+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Measurements+CoreDataProperties.swift
 //  WeatherStation
 //
-//  Created by Kostiantyn Dovbush on 12/01/2017.
+//  Created by Kostiantyn Dovbush on 1/21/17.
 //  Copyright © 2017 Kostiantyn Dovbush. All rights reserved.
 //
 
@@ -18,8 +18,8 @@ extension Measurements {
 
     @NSManaged public var createdAt: NSDate
     @NSManaged public var heatIndex: Double
-    @NSManaged public var humidity: Double
-    @NSManaged public var rainAnalog: Double
+    @NSManaged public var humidity: Int32
+    @NSManaged public var rainAnalog: Int32
     @NSManaged public var rainDigital: Bool
     @NSManaged public var temperature: Double
     @NSManaged public var station: Station?
